+++ src/error.rs
@@ -86,7 +86,7 @@ pub enum TocError {
 	/// # CDDB Decode.
 	CddbDecode,
 
-	#[cfg(all(feature = "base64", feature = "sha1"))]
+	#[cfg(feature = "sha1")]
 	/// # SHA1/Base64 Decode.
 	ShaB64Decode,
 }
@@ -108,7 +108,7 @@ impl fmt::Display for TocError {
 
 			#[cfg(feature = "accuraterip")] Self::AccurateRipDecode => f.write_str("Invalid AccurateRip ID string."),
 			#[cfg(feature = "cddb")] Self::CddbDecode => f.write_str("Invalid CDDB ID string."),
-			#[cfg(all(feature = "base64", feature = "sha1"))] Self::ShaB64Decode => f.write_str("Invalid sha/base64 ID string."),
+			#[cfg(feature = "sha1")] Self::ShaB64Decode => f.write_str("Invalid sha/base64 ID string."),
 		}
 	}
 }