+++ src/musicbrainz.rs
@@ -9,6 +9,10 @@ use crate::{
 
 
 
+const CHUNK_SIZE: usize = 4;
+
+
+
 impl Toc {
 	#[allow(clippy::cast_possible_truncation, clippy::missing_panics_doc)]
 	#[cfg_attr(docsrs, doc(cfg(feature = "musicbrainz")))]
@@ -32,24 +36,50 @@ impl Toc {
 	pub fn musicbrainz_id(&self) -> ShaB64 {
 		use sha1::Digest;
 		let mut sha = sha1::Sha1::new();
-		let mut buf: [u8; 8] = [b'0', b'1', b'0', b'0', b'0', b'0', b'0', b'0'];
+		let mut src = [b'0'; CHUNK_SIZE * 4]; // Four raw u32s.
+		let mut dst: [u8; CHUNK_SIZE * 8] = [
+			b'0', b'1', b'0', b'0', b'0', b'0', b'0', b'0',
+			b'0', b'0', b'0', b'0', b'0', b'0', b'0', b'0',
+			b'0', b'0', b'0', b'0', b'0', b'0', b'0', b'0',
+			b'0', b'0', b'0', b'0', b'0', b'0', b'0', b'0',
+		]; // Four hexed u32s.
 
-		// Start with "01" and the audio track count.
-		faster_hex::hex_encode(&[self.audio_len() as u8], &mut buf[2..4]).unwrap();
-		buf[2..4].make_ascii_uppercase();
-		sha.update(&buf[..4]);
+		// Start with "01", the audio track count, and leadout.
+		faster_hex::hex_encode_fallback(&[self.audio_len() as u8], &mut dst[2..4]);
+		faster_hex::hex_encode_fallback(self.audio_leadout().to_be_bytes().as_slice(), &mut dst[4..12]);
+		dst[2..12].make_ascii_uppercase();
+		sha.update(&dst[..12]);
 
-		// Add the audio leadout.
-		faster_hex::hex_encode(self.audio_leadout().to_be_bytes().as_slice(), &mut buf).unwrap();
-		buf.make_ascii_uppercase();
-		sha.update(buf.as_slice());
-
-		// Now the audio starts.
+		// Process the sector positions in batches of four to leverage SSE hex
+		// optimizations.
 		let sectors = self.audio_sectors();
-		for &v in sectors {
-			faster_hex::hex_encode(v.to_be_bytes().as_slice(), &mut buf).unwrap();
-			buf.make_ascii_uppercase();
-			sha.update(buf.as_slice());
+		let len = sectors.len();
+		let rem = len % CHUNK_SIZE;
+		for v in sectors.chunks_exact(CHUNK_SIZE) {
+			// Copy the values to the source buffer.
+			for (s_chunk, v) in src.chunks_exact_mut(4).zip(v) {
+				s_chunk.copy_from_slice(v.to_be_bytes().as_slice());
+			}
+
+			// Encode and hash, en masse.
+			faster_hex::hex_encode(src.as_slice(), &mut dst).unwrap();
+			dst.make_ascii_uppercase();
+			sha.update(dst.as_slice());
+		}
+
+		// Handle the remaining sectors, if any,
+		if rem != 0 {
+			// Copy the values to the source buffer.
+			for (s_chunk, v) in src.chunks_exact_mut(4).zip(sectors[len - rem..].iter()) {
+				s_chunk.copy_from_slice(v.to_be_bytes().as_slice());
+			}
+
+			// Encode and hash, en masse.
+			let src_to = rem * 4;
+			let dst2 = &mut dst[..src_to * 2];
+			faster_hex::hex_encode_fallback(&src[..src_to], dst2);
+			dst2.make_ascii_uppercase();
+			sha.update(dst2);
 		}
 
 		// Pad with zeroes.
@@ -69,7 +99,7 @@ mod tests {
 
 	#[test]
 	fn t_musicbrainz() {
-		for (t, c) in [
+		for (t, id) in [
 			(
 				"18+B6+3CE3+7C6F+B2BD+E47F+1121C+15865+175E0+1AED9+1E159+20BF9+235FC+259EF+2826E+29B62+2ED67+311B1+3396B+36ACB+3916B+3BB75+3D60A+40AA6+422FE+48B68+4E4CB",
 				"eLuEIkHsua.iJpetabxqYM9SIbk-",
@@ -96,7 +126,13 @@ mod tests {
 			),
 		] {
 			let toc = Toc::from_cdtoc(t).expect("Invalid TOC");
-			assert_eq!(toc.musicbrainz_id().to_string(), c);
+			let mb_id = toc.musicbrainz_id();
+			assert_eq!(mb_id.to_string(), id);
+
+			// Test decoding three ways.
+			assert_eq!(ShaB64::decode(id), Ok(mb_id));
+			assert_eq!(ShaB64::try_from(id), Ok(mb_id));
+			assert_eq!(id.parse::<ShaB64>(), Ok(mb_id));
 		}
 	}
 }